move some of common into tokens.py
