+++ netam/models.py
@@ -13,13 +13,15 @@ from netam.hit_class import apply_multihit_correction
 from netam import sequences
 from netam.sequences import MAX_AA_TOKEN_IDX
 from netam.common import (
-    aa_idx_tensor_of_str_ambig,
-    PositionalEncoding,
+    chunk_function,
+    assume_single_sequence_is_heavy_chain,
+)
+from netam.tokens import (
     generate_kmers,
     aa_mask_tensor_of,
     encode_sequences,
-    chunk_function,
-    assume_single_sequence_is_heavy_chain,
+    aa_idx_tensor_of_str_ambig,
+    PositionalEncoding,
 )
 
 from netam.sequences import set_wt_to_nan
